Change `ProgressBar` to be compatible with OpenMP
`ProgressBar` was initially used directly in the initialization
statements of `for`-loops. However, OpenMP requires that `for`-loops
have canonical form, which as of the OpenMP 5.2 specification, does not
allow types other than integer, pointer, or random-access iterators to
be used in the init-statement of a `for`-loop. Furthermore,
parallelizing a `for`-loop with a `ProgressBar` as its loop variable
would mess up the `ProgressBar`'s internal percentage tracker, because
it depends on the current value of the loop variable, which is different
for different threads.