Remove `Camera::render_to()`; can't parallelize it
The difference between render_to(ray_color, "output.ppm") and
render(ray_color).send_as_ppm("output.ppm") was that `render_to` used
the `ImagePPMStream` type to output the image, which saves storage by
not storing the pixels of the image, instead requiring the pixel values
to be progressively given through `ImagePPMStream::add()` in order from
top to bottom then from left to right. However, this does not play well
with multi-threading; currently, different threads will render different
rows at once, meaning `ImagePPMStream` will output the pixels in the
wrong order. The only options are either to parallelize on the very
inner loop (that loops over `samples_per_pixel`), which would likely
increase running time due to the overhead of having to creating a thread
pool for every single pixel in the image, or to parallelize but to store
the pixels and output them all at the end. I decided to go with the
latter option, which would mean that `render_to` saves no storage over
`render` anyways, and so we might as well delete it and leave only
`Camera::render()`.