Fix thread unsafety in rand_double()
Before this commit, calling rand_double() concurrently was not thread
safe, because there was only a single static std::mt19937, and
having multiple threads use it to generate random numbers (having
multiple threads try to advance the PRNG state at once) is not
thread-safe. To fix this, we apply the C++ thread_local keyword to
the std::mt19937. This means that each thread will have its own
instance of the std::mt19937 (seeded differently), which makes
rand_double() thread safe.