Optimize hit_info, small corrections
hit_info originally stored the material of the object hit as a
std::shared_ptr<Material>. This is unnecessary because objects of type
hit_info only exist as an intermediate product inside Camera::render(),
which does not modify the materials of objects in the Scene passed to
it, and so the hit_info's generated do not need to share ownership of
the materials. Thus, we change the material field to a const Material*.
This is a significant optimization because it makes the destructor of
hit_info trivial, which makes the destructor of std::optional<hit_info>
trivial, which speeds up calls to hit_by().