Remove default argument in Scene::hit_by()
The parameter ray_times in Scene::hit_by() previously had a default
argument; namely, Interval::nonnegative(). However, this is a poor
default, because checking for intersections at just all non-negative
times leads to shadow acne (the fix is to allow intersection times
greater than some small positive constant), and so we remove it. The
same default argument for ray_times in Sphere::hit_by() was removed
in commit 784daf5.