Update Camera; use std::optional, add lookat field
Remove the use of sentinel values for the Camera fields horizontal_fov,
vertical_fov, and focus_dist by making them all std::optional's. Also,
add the concept of a "lookat" point for the Camera (a point which the
camera always points toward, no matter what the camera center is set
to) and its corresponding setter function set_camera_lookat. The
previous function with the same name is renamed to
set_camera_direction_towards; see the code for comments.