Use dynamic thread scheduling in Camera::render()
For the render loop in Camera::render(), we change from using static
scheduling to using dynamic scheduling, with a block size equal to the
larger of the number of rows divided by 1024, and 1. This prevents
threads from running out of work too early and idling, while also
ensuring that the threads don't wait too long to be assigned new rows to
render (the total number of times that threads wait to be assigned a new
chunk of iterations is O(1024)). This results in a ~10% performance
improvement in the render time for images with larger numbers of rows.