Add diffuse lights/custom fixed random seeding
Support for diffuse lights (lights that uniformly emit the same color in
all directions) is added through the DiffuseLight material. We also add
SeedSeqGenerator in rand_util.h, which allows the user to specify a seed
to generate the sequence of seeds used by the different threads' RNGs;
this allows the user to work on the same scene between different runs
of the program. Finally, note that the diffuse lights support colors
(DiffuseLight has an intrinsic_color field), but it doesn't work
completely correctly; all the lights themselves show up as white instead
of their respectively colors, because their calculated color (which is
their intensity multiplied by their intrinsic color) has RGB magnitudes
all exceeding 1, and so their colors show up as white. This can be fixed
through tone mapping, which will be implemented in a future commit.