Fix RGB::operator+, rerender image
RGB::operator+ was introduced in commit 1bd6dd9, where it was
implemented incorrectly; the green and blue components were mistakenly
switched in the returned RGB, resulting in incorrect colors in the
rendered image "rtow_final_lights_without_tone_mapping.png". We fix
the implementation of RGB::operator+ and rerender the image.