Add tone mapping; display light colors correctly
Commit 1bd6dd9 added basic colored diffuse lights. However, when
rendering colored lights with large intensities, the lights themselves
just become completely white, their intrinsic color only evident from
the color they cast on the surrounding area (as seen in the rendered
image rtow_final_lights_without_tone_mapping.png). This happens because
the emitted color from a diffuse light is just the intrinsic color
multiplied by the intensity. Thus, when the intensity is high enough,
the emitted color will have RGB magnitudes all greater than 1, and will
thus show up as white in the final image, even though the ratio between
the R/G/B components still do correctly represent the intrinsic color.
This is a well-known problem in HDR (High Dynamic Range) rendering, and
is solved through Tone Mapping: mapping the set of colors including
those with RGB magnitudes all greater than 1, to the set of colors with
RGB magnitudes from 0 to 1 inclusive. Tone mapping is done through a
tone mapping operator, a very popular example of which is the Reinhard
operator. In this commit, we implement the Reinhard operator in
RGB::as_string(); it simply divides each of the R/G/B components by one
plus the luminance of the RGB color, which is calculated by another
well-known formula in computer graphics. See rgb.h for the code.
Finally, we rerender the scene with colored spherical lights in the
and store it in the new file "rtow_final_lights_with_tone_mapping.png";
now, the colors all show up correctly.