Correct focus distance for last render, rerender
The previous render in images/millions_of_spheres_with_lights.png had
the focus distance set to 10. This was a mistake; it was intended for
the origin to be in perfect focus. We set the focus distance to be
approximately the distance from the camera center to the origin (51),
and rerender the image.