Make ground flatter in rtow_final_lights images
Spheres are the only primitive supported as of now. As a result, the
"ground" in every rendered image is just a sphere with a large radius.
In rtow_final_lights_with/without_tone_mapping.png, however, the ground
sphere's radius was too small, resulting in the ground curving below the
small spheres in the image, causing their shadows to appear detached. We
increase the sphere's radius from 1 thousand to 1 million to fix this,
and rerender both images.