Add top-level Makefile
Commit 2f6bfc9 moved all source code as well as the Makefile into the
src/ directory. While improving organization, this also resulted in
"no makefile found" errors occurring when trying to build from the
top-level directory, because the Makefile had been moved out of the
top-level directory into src/. Now, we resolve this by adding a
top-level Makefile that simply delegates to the Makefile in src/.
According to a Stack Overflow answer linked in the new top-level
Makefile, this is the "traditional way" of solving this problem.