Optimize BVH/AABB/Camera and fix Makefile
We implement four significant optimizations to our ray-geometry
intersection tests. Firstly, we flatten the BVH tree into an array,
which improves cache, memory usage, and overall system performance.
Secondly, we replace the recursive algorithm for traversing the BVH
during ray-geometry intersection tests with an iterative stack-based
DFS. Thirdly, we introduce and incorporate AABB::is_hit_by_optimized()
into the BVH intersection routine; this is a specialized version of
AABB::is_hit_by() that has extra parameters that correspond to
precomputed information which is constant for a given ray, allowing us
to avoid recomputation of values. Finally, during ray-geometry
intersection tests, whenever an interior BVH node is traversed (which
requires us to visit both children of that node), we now choose to visit
the closer child node first (as determined by which axis its parent
interior node was split by, as well as the direction of the ray); this
allows the time interval to be shrunk as much as possible by the time we
visit the further node, increasing the probability that its bounding box
will not be hit by the ray in the desired time interval, and therefore
decreasing the total number of BVH nodes visited by ray-geometry
intersection tests. In the end, these four optimizations improve overall
performance by 40-50%, resulting in runtimes that are 65-75% of the
original.

Additionally, we use C++20 concepts to guarantee that static dispatch
will be used in functions which used to take a general Hittable
parameter, such as Camera::render(), Camera::ray_color(), and
BVH::BVH().

Finally, we use the PRECIOUS target in Makefile to fix a bug where after
a make clean, make would need to be called twice in order for the .dbg.o
and the .rel.o files to generate (the .dbg.o and .rel.o files were being
automatically deleted by make after the first command; setting the
PRECIOUS target to .dbg.o and .rel.o files allows us to prevent this
deletion from occurring).