Remove directories previous_parts and .vscode
This commit removes two top-level directories: previous_parts, and
.vscode. The previous_parts directory was used to record the progression
of the main.cpp file as I worked through "Ray Tracing in One Weekend".
However, all such main.cpp files are already available in the commit
history, which also contains the previous revisions of the header files
that any reader would need to know to be able to interpret those
previous versions of main.cpp in the first place. Thus, there is no
benefit to keeping the previous_parts directory, and so we remove it.
The .vscode directory is removed for now because it contains only
generic configuration and settings files specific to my own personal
usage of Visual Studio Code; those files are not necessary for other
users to build nor run the code.