Improve ray-parallelogram intersections
See parallelogram.h and Parallelogram::hit_by() for more info. As a
summary, a poor choice for the normal vector used in the
ray-parallelogram intersection routine resulted in parallelograms with
large coordinates rejecting all rays and failing to render as a result.
We resolve this issue with a better choice of normal vector.