Fix and improve SeedSeqGenerator
We improve SeedSeqGenerator, the class used to generate random seeds for
the thread_local RNGs used in the rand_double() function. First, we make
SeedSeqGenerator a singleton class, which avoids confusion about which
instance should be used (because a singleton class only has a single
instance). Secondly, we fix the thread unsafety that was previously in
SeedSeqGenerator::operator() (now renamed to next_seed()). Thirdly, we
add output to notify the user of the seed being used, or, if no seed was
provided, we output instructions on how to set a random seed. Finally,
we add comments that explain the importance of the specific constants
used in the LCG generators used in both rand_double() and in
SeedSeqGenerator::next_seed().

However, because we change the random generation algorithm used to
generate seeds within SeedSeqGenerator, previous renders with fixed
random seeds change. For now, I have adjusted the seed for
rtow_final_lights_with_tone_mapping() so that the same scene is still
rendered. In general, for scenes that are generated by a single thread
(so the scene is determined by the very first seed returned from
SeedSeqGenerator::seed_by()), it is possible to solve for the new seed
that generates the same first seed as the original seed, because
seed_by() uses a LCG internally.